book_review/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
sqlparse==0.5.0
dj-database-url==2.2.0 
psycopg2-binary==2.9.9